third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-console/debugger-command-line-api-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that inspect() command line api works while on breakpoint.

Set timer for test function.
Script execution paused.
WebInspector.inspect called with: p#p1
WebInspector.inspect's hints are: []
inspect($('#p1')) = 
Selected node id: 'p1'.
Script execution resumed.
